Fixed fixture templates, reporting and plotting
Continuation of feature porting to JSON.

Fixtures can now be created from existing templates in the content
directory.

Reports can be created but template searching is not yet functional
(waiting on a rewrite of data.py) so the directory must be manually
changed in the plotter.py file.

Plots now get generated at their proper size by setting the viewport so
1px = 1mm.